added main_roating_sorted_numbers
Fixed of the problem: I just want to add one comment. monotonically_increasing_id does not guarantee consecutive numbers it only guarantees that they are increasing. While it works for small datasets since you are probably not distributing it, it will not work for large datasets. Try it on a large dataset and compare the last id to count. This will make the part of the code break where you are joining the two columns break for very large input..